sqweasdq

+++ database/migrations/2022_05_07_105540_create_telegram_users_table.php
@@ -16,7 +16,7 @@ class CreateTelegramUsersTable extends Migration
     {
         Schema::create('telegram_users', function (Blueprint $table) {
             $table->id();
-            $table->integer('telegram_id')->unique();
+            $table->bigInteger('telegram_id')->unique();
             $table->string('number')->unique();
             $table->string('last_name');
             $table->string('first_name');
